refactor and add logfmt
