+++ Dockerfile
@@ -4,4 +4,7 @@ RUN apk add --no-cache git curl
 WORKDIR "/go"
 
 RUN curl https://glide.sh/get | sh
-RUN go get golang.org/x/tools/cmd/godoc
+RUN go get \
+    golang.org/x/tools/cmd/godoc \
+    cmd/gofmt \
+    golang.org/x/tools/cmd/goimports
