+++ logging/statsd_test.go
@@ -11,94 +11,102 @@
 package logging
 
 import (
-    "testing"
-    "github.com/stretchr/testify/assert"
-    "github.com/DataDog/datadog-go/statsd"
-    "../nettest"
-    "time"
-    "os"
-    "net/http"
-    "net/url"
+	"net"
+	"os"
+	"testing"
+
+	"../nettest"
+	"github.com/stretchr/testify/assert"
 )
 
-func getDuration(t *testing.T, dur string) (duration time.Duration) {
-    duration, err := time.ParseDuration(dur)
-    if err != nil {
-        t.Fatal(err)
-    }
-    return
-}
-
 func TestStatsdLogging(t *testing.T) {
-    cases := map[string]struct{
-        request *http.Request
-        timestamp time.Time
-        duration time.Duration
-        expected []string
-    }{
-        "basic": {
-            newRequest("GET", "http://example.com"),
-            time.Now().UTC(),
-            getDuration(t, "0.302s"),
-            []string{
-                "service.logging.live.request.response_time:302.000000|ms|#test,endpoint:/,statusCode:200,method:GET,protocol:HTTP/1.1",
-                "service.logging.live.request.count:1|c|#test,endpoint:/,statusCode:200,method:GET,protocol:HTTP/1.1",
-            },
-        },
-        "post path": {
-            newRequest("POST", "http://example.com/path/here"),
-            time.Now().UTC(),
-            getDuration(t, "0.102s"),
-            []string{
-                "service.logging.live.request.response_time:102.000000|ms|#test,endpoint:/path/here,statusCode:200,method:POST,protocol:HTTP/1.1",
-                "service.logging.live.request.count:1|c|#test,endpoint:/path/here,statusCode:200,method:POST,protocol:HTTP/1.1",
-            },
-        },
-        "strips params off method": {
-            newRequest("GET", "http://example.com/token/1/test?apid=1&thing=2"),
-            time.Now().UTC(),
-            getDuration(t, "0.927s"),
-            []string{
-                "service.logging.live.request.response_time:927.000000|ms|#test,endpoint:/token/1/test,statusCode:200,method:GET,protocol:HTTP/1.1",
-                "service.logging.live.request.count:1|c|#test,endpoint:/token/1/test,statusCode:200,method:GET,protocol:HTTP/1.1",
-            },
-        },
-        "connect http2 test": {
-            &http.Request{
-                Method:     "CONNECT",
-                Proto:      "HTTP/2.0",
-                ProtoMajor: 2,
-                ProtoMinor: 0,
-                URL:        &url.URL{Host: "www.example.com:443"},
-                Host:       "www.example.com:443",
-                RemoteAddr: "192.168.100.5",
-            },
-            time.Now().UTC(),
-            getDuration(t, "0.927s"),
-            []string{
-                "service.logging.live.request.response_time:927.000000|ms|#test,endpoint:www.example.com:443,statusCode:200,method:CONNECT,protocol:HTTP/2.0",
-                "service.logging.live.request.count:1|c|#test,endpoint:www.example.com:443,statusCode:200,method:CONNECT,protocol:HTTP/2.0",
-            },
-        },
-    }
+	tests := map[string]struct {
+		metric   string
+		value    float64
+		tags     []string
+		expected string
+	}{
+		"base test": {
+			"request.count",
+			1,
+			[]string{},
+			"request.count:1|c",
+		},
+		"tags test": {
+			"request.tags",
+			1,
+			[]string{"tag1", "tag2:value"},
+			"request.tags:1|c|#tag1,tag2:value",
+		},
+	}
 
-    done := make(chan string)
-    addr, sock, srvWg := nettest.CreateServer(t, "udp", "localhost:", done)
-    defer srvWg.Wait()
-    defer os.Remove(addr.String())
-    defer sock.Close()
+	done := make(chan string)
+	addr, sock, srvWg := nettest.CreateServer(t, "udp", "localhost:", done)
+	defer srvWg.Wait()
+	defer os.Remove(addr.String())
+	defer sock.Close()
 
-	client, err := statsd.New(addr.String())
+	host, port, err := net.SplitHostPort(addr.String())
 	if err != nil {
 		t.Fatal(err)
 	}
-    client.Tags = append(client.Tags, "test")
-    client.Namespace = "service.logging.live."
+	os.Setenv("STATSD_HOST", host)
+	os.Setenv("STATSD_PORT", port)
 
-    for k, tc := range cases {
-        writeStatsdLog(client, tc.request, *tc.request.URL, tc.timestamp, tc.duration, http.StatusOK, 100)
-        for _, message := range tc.expected {
-            assert.Equal(t, message, <-done, "test: %s", k)
-        }
-    }
+	client, err := GetStatsdFromEnv()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	for k, tc := range tests {
+		client.Incr(tc.metric, tc.tags, tc.value)
+		assert.Equal(t, tc.expected, <-done, "test: %s", k)
+	}
+}
+
+func TestStatsdLoggingWithNamespaceAndTags(t *testing.T) {
+	tests := map[string]struct {
+		metric   string
+		value    float64
+		tags     []string
+		expected string
+	}{
+		"base test": {
+			"request.count",
+			1,
+			[]string{},
+			"service.request.count:1|c|#tag1:value,tag2",
+		},
+		"extra tags": {
+			"request.count",
+			1,
+			[]string{"tag3,tag4:thing"},
+			"service.request.count:1|c|#tag1:value,tag2,tag3,tag4:thing",
+		},
+	}
+
+	done := make(chan string)
+	addr, sock, srvWg := nettest.CreateServer(t, "udp", "localhost:", done)
+	defer srvWg.Wait()
+	defer os.Remove(addr.String())
+	defer sock.Close()
+
+	host, port, err := net.SplitHostPort(addr.String())
+	if err != nil {
+		t.Fatal(err)
+	}
+	os.Setenv("STATSD_HOST", host)
+	os.Setenv("STATSD_PORT", port)
+	os.Setenv("STATSD_NAMESPACE", "service.")
+	os.Setenv("STATSD_TAGS", "tag1:value,tag2")
+
+	client, err := GetStatsdFromEnv()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	for k, tc := range tests {
+		client.Incr(tc.metric, tc.tags, tc.value)
+		assert.Equal(t, tc.expected, <-done, "test: %s", k)
+	}
 }
